requirements.txt
faiss-cpu
fastapi
langchain
langchain-community
langchain-openai
pymupdf
pytesseract
python-dotenv
uvicorn
requests
streamlit
streamlit-pdf-viewer
